Build the static prototype on the Router object
Lines 26-36: static prototype takes a path as an argument, reads thats path.
For each of the files in the specified directory, sets this.routes to url of current file, return servedFile as a function call.

Needs revising. Unsure of how to create working GET requests for multiple files if ```var servedFile``` can only hold one route at a time, and is returned on line 32.